Move functionality out of main.cc
