+++ kvrocks.conf
@@ -127,6 +127,13 @@ db-name change.me.db
 # Note that you must specify a directory here, not a file name.
 dir /tmp/kvrocks
 
+# Optional: A snapshot directory
+#
+# If specified, Kvrocks will snapshot the DB at [dir] to the specified [snapshot-dir], 
+# and start a read-only server from the snapshot.
+# 
+# snapshot-dir /tmp/kvrocks-snapshot-changeme
+
 # You can configure where to store your server logs by the log-dir.
 # If you don't specify one, we will use the above `dir` as our default log directory.
 # We also can send logs to stdout/stderr is as simple as:
