+++ src/config/config.cc
@@ -24,13 +24,11 @@
 #include <rocksdb/env.h>
 #include <strings.h>
 
-#include <algorithm>
 #include <cstdint>
 #include <cstring>
 #include <fstream>
 #include <iostream>
 #include <iterator>
-#include <limits>
 #include <string>
 #include <utility>
 #include <vector>
@@ -191,7 +189,6 @@ Config::Config() {
       {"db-name", true, new StringField(&db_name, "change.me.db")},
       {"dir", true, new StringField(&dir, kDefaultDir)},
       {"snapshot-dir", false, new StringField(&snapshot_dir, "")},
-      {"keep-snapshot", false, new YesNoField(&keep_snapshot, false)},
       {"backup-dir", false, new StringField(&backup_dir, kDefaultBackupDir)},
       {"log-dir", true, new StringField(&log_dir, "")},
       {"log-level", false, new EnumField<int>(&log_level, log_levels, google::INFO)},
